+++ include/z64actor.h
@@ -97,18 +97,18 @@ typedef struct {
 } ActorOverlay; // size = 0x20
 
 typedef struct {
-    u8 table[32];
+    u16 table[32];
 } DamageTable;
 
 typedef struct {
-    /* 0x00 */ u8 health;
+    /* 0x00 */ u16 health;
     /* 0x02 */ s16 cylRadius;
     /* 0x04 */ s16 cylHeight;
     /* 0x06 */ u8 mass;
 } CollisionCheckInfoInit;
 
 typedef struct {
-    /* 0x00 */ u8 health;
+    /* 0x00 */ u16 health;
     /* 0x02 */ s16 cylRadius;
     /* 0x04 */ s16 cylHeight;
     /* 0x06 */ s16 cylYShift;
@@ -122,8 +122,8 @@ typedef struct {
     /* 0x12 */ s16 cylHeight; // Used for various purposes
     /* 0x14 */ s16 cylYShift; // Unused. Purpose inferred from Cylinder16 and CollisionCheck_CylSideVsLineSeg
     /* 0x16 */ u8 mass; // Used to compute displacement for OC collisions
-    /* 0x17 */ u8 health; // Note: some actors may use their own health variable instead of this one
-    /* 0x18 */ u8 damage; // Amount to decrement health by
+    /* 0x17 */ u16 health; // Note: some actors may use their own health variable instead of this one
+    /* 0x18 */ u16 damage; // Amount to decrement health by
     /* 0x19 */ u8 damageEffect; // Stores what effect should occur when hit by a weapon
     /* 0x1A */ u8 atHitEffect; // Stores what effect should occur when AT connects with an AC
     /* 0x1B */ u8 acHitEffect; // Stores what effect should occur when AC is touched by an AT
